realziar despliegue de actividades

+++ ChronosSuite/Models/Location.cs
@@ -5,10 +5,10 @@ namespace ChronosSuite.Models;
 
 public partial class Location
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
 
     public string Name { get; set; } = null!;
-    
+
     public string? Description { get; set; }
 
     public virtual ICollection<VisitRecord> VisitRecords { get; set; } = new List<VisitRecord>();
